Use default user-agent and ca_data in op_create_http_client
Also use a single create_http_client implementation (removing runtime::http_util)

Fixes #9827
