+++ runtime/ops/fetch.rs
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
-use crate::http_util;
 use crate::permissions::Permissions;
 use deno_fetch::reqwest;
+use deno_fetch::HttpClientDefaults;
 
 pub fn init(
   rt: &mut deno_core::JsRuntime,
@@ -12,7 +12,11 @@ pub fn init(
     let op_state = rt.op_state();
     let mut state = op_state.borrow_mut();
     state.put::<reqwest::Client>({
-      http_util::create_http_client(user_agent, ca_data).unwrap()
+      deno_fetch::create_http_client(user_agent.clone(), ca_data.clone()).unwrap()
+    });
+    state.put::<HttpClientDefaults>(HttpClientDefaults {
+      ca_data: ca_data.clone(),
+      user_agent: user_agent.clone(),
     });
   }
   super::reg_json_sync(rt, "op_fetch", deno_fetch::op_fetch::<Permissions>);
